+++ dragonchain/webserver/lib/interchain_utest.py
@@ -19,6 +19,7 @@ import unittest
 from unittest.mock import patch, MagicMock
 
 from dragonchain import test_env  # noqa: F401
+from dragonchain import exceptions
 from dragonchain.webserver.lib import interchain
 
 
@@ -95,21 +96,37 @@ class TestInterchain(unittest.TestCase):
 
     @patch("dragonchain.lib.dto.btc.new_from_user_input")
     @patch("dragonchain.lib.dao.interchain_dao.save_interchain_client")
+    @patch("dragonchain.lib.dao.interchain_dao.does_interchain_exist", return_value=False)
     @patch("dragonchain.webserver.lib.interchain._get_output_dto_v1", return_value="blah")
-    def test_create_interchain_bitcoin(self, mock_get_output, mock_save_client, mock_new_from_user):
+    def test_create_interchain_bitcoin(self, mock_get_output, mock_exist, mock_save_client, mock_new_from_user):
         self.assertEqual(interchain.create_bitcoin_interchain_v1({"thing": "ok"}), "blah")
         mock_new_from_user.assert_called_once_with({"thing": "ok"})
         mock_save_client.assert_called_once_with(mock_new_from_user.return_value)
         mock_get_output.assert_called_once_with(mock_new_from_user.return_value)
 
+    @patch("dragonchain.lib.dto.btc.new_from_user_input")
+    @patch("dragonchain.lib.dao.interchain_dao.does_interchain_exist", return_value=True)
+    def test_create_interchain_bitcoin_throws_if_already_exists(self, mock_exist, mock_new_from_user):
+        self.assertRaises(exceptions.InterchainConflict, interchain.create_bitcoin_interchain_v1, {})
+        mock_new_from_user.assert_called_once_with({})
+        mock_exist.assert_called_once()
+
     @patch("dragonchain.lib.dto.eth.new_from_user_input")
     @patch("dragonchain.lib.dao.interchain_dao.save_interchain_client")
+    @patch("dragonchain.lib.dao.interchain_dao.does_interchain_exist", return_value=False)
     @patch("dragonchain.webserver.lib.interchain._get_output_dto_v1", return_value="blah")
-    def test_create_interchain_ethereum(self, mock_get_output, mock_save_client, mock_new_from_user):
+    def test_create_interchain_ethereum(self, mock_get_output, mock_exist, mock_save_client, mock_new_from_user):
         self.assertEqual(interchain.create_ethereum_interchain_v1({"thing": "ok"}), "blah")
         mock_new_from_user.assert_called_once_with({"thing": "ok"})
         mock_save_client.assert_called_once_with(mock_new_from_user.return_value)
         mock_get_output.assert_called_once_with(mock_new_from_user.return_value)
+
+    @patch("dragonchain.lib.dto.eth.new_from_user_input")
+    @patch("dragonchain.lib.dao.interchain_dao.does_interchain_exist", return_value=True)
+    def test_create_interchain_ethereum_throws_if_already_exists(self, mock_exist, mock_new_from_user):
+        self.assertRaises(exceptions.InterchainConflict, interchain.create_ethereum_interchain_v1, {})
+        mock_new_from_user.assert_called_once_with({})
+        mock_exist.assert_called_once()
 
     @patch("dragonchain.webserver.lib.interchain.create_bitcoin_interchain_v1", return_value="cool")
     @patch("dragonchain.lib.dao.interchain_dao.get_interchain_client")