+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Bugs:**
   - Fix a bug where getting the cached list of verifications for a block would always fail
+  - Fix a bug where existing interchain networks could be overwritten if trying to create a new network with the same blockchain and name
 - **Packaging:**
   - Update aioredis, boto3, web3, and aiohttp dependencies